bridge/pyproject.toml
[project]
name = "bridge"
version = "0.1.0"
description = "HandOver AI - Bridge Application"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pywebview>=5.0",
    "typing-extensions>=4.0",
    "psutil>=5.9.0",
    # AI dependencies
    "openai>=1.0.0",
    "chromadb>=0.4.0",
    "sentence-transformers>=2.2.0",
    "rank-bm25>=0.2.2",
    "langgraph>=0.0.20",
    # Document Parsers
    "pypdf>=3.0.0",
    "python-docx>=1.2.0",
    "olefile>=0.46",
    "openpyxl>=3.0.0",
    "xlrd>=2.0.0",
    "pandas>=1.5.0",
    "pyhwp>=0.1b15",
    "numpy>=1.24.0",
    "fastapi>=0.128.0",
    "pypdf2>=3.0.1",
    "pypdf>=3.0.0",
    "uvicorn>=0.40.0",
    "opentelemetry-api>=1.39.1",
    "opentelemetry-instrumentation>=0.60b1",
    "opentelemetry-sdk>=1.39.1",
    "jaraco-text>=4.0.0",
    "jaraco-classes>=3.4.0",
    "jaraco-context>=6.1.0",
    "platformdirs>=4.5.1",
    "opentelemetry-instrumentation-fastapi>=0.60b1",
]

[dependency-groups]
dev = [
    "pyinstaller>=6.18.0",
    "pillow>=10.0.0",
]
